backend/phishing/requirements.txt

dnspython==2.4.2
python-whois==0.8.0
strsimpy==0.2.1
flask==3.0.0
requests==2.31.0
